+++ src/api.js
@@ -27,7 +27,8 @@ async function listSpaces (address, serverUrl = PROFILE_SERVER_URL) {
   }
 }
 
-async function getSpace (address, name, serverUrl = PROFILE_SERVER_URL, { metadata }) {
+async function getSpace (address, name, serverUrl = PROFILE_SERVER_URL, { metadata, blocklist } = {}) {
+  if (blocklist && blocklist(address)) throw new Error(`user with ${address} is blocked`)
   let url = `${serverUrl}/space`
 
   try {
@@ -54,18 +55,47 @@ async function getSpace (address, name, serverUrl = PROFILE_SERVER_URL, { metada
   }
 }
 
-async function getThread (space, name, serverUrl = PROFILE_SERVER_URL) {
-  return new Promise(async (resolve, reject) => {
-    try {
-      const res = await utils.fetchJson(serverUrl + `/thread?space=${encodeURIComponent(space)}&name=${encodeURIComponent(name)}`)
-      resolve(res)
-    } catch (err) {
-      reject(err)
-    }
-  })
+// TODO consumes address now, could also give root DID to get space DID
+async function getSpaceDID (address, space, opts = {}) {
+  const conf = await getConfig(address, opts)
+  if (!conf.spaces[space] || !conf.spaces[space].DID) throw new Error(`Could not find appropriate DID for address ${address}`)
+  return conf.spaces[space].DID
 }
 
-async function getProfile (address, serverUrl = PROFILE_SERVER_URL, { metadata }) {
+async function getThread (space, name, firstModerator, members, opts = {}) {
+  const serverUrl = opts.profileServer || PROFILE_SERVER_URL
+  if (firstModerator.startsWith('0x')) {
+    firstModerator = await getSpaceDID(firstModerator, space, opts)
+  }
+  try {
+    let url = `${serverUrl}/thread?space=${encodeURIComponent(space)}&name=${encodeURIComponent(name)}`
+    url += `&mod=${encodeURIComponent(firstModerator)}&members=${encodeURIComponent(members)}`
+    return await utils.fetchJson(url)
+  } catch (err) {
+    throw new Error(err)
+  }
+}
+
+async function getThreadByAddress (address, opts = {}) {
+  const serverUrl = opts.profileServer || PROFILE_SERVER_URL
+  try {
+    return await utils.fetchJson(`${serverUrl}/thread?address=${encodeURIComponent(address)}`)
+  } catch (err) {
+    throw new Error(err)
+  }
+}
+
+async function getConfig (address, opts = {}) {
+  const serverUrl = opts.profileServer || PROFILE_SERVER_URL
+  try {
+    return await utils.fetchJson(`${serverUrl}/config?address=${encodeURIComponent(address)}`)
+  } catch (err) {
+    throw new Error(err)
+  }
+}
+
+async function getProfile (address, serverUrl = PROFILE_SERVER_URL, { metadata, blocklist } = {}) {
+  if (blocklist && blocklist(address)) throw new Error(`user with ${address} is blocked`)
   let url = `${serverUrl}/profile`
 
   try {
@@ -112,7 +142,7 @@ async function profileGraphQL (query, opts = {}) {
 }
 
 async function getVerifiedAccounts (profile) {
-  let verifs = {}
+  const verifs = {}
   try {
     const did = await verifier.verifyDID(profile.proof_did)
 
@@ -145,4 +175,4 @@ async function getVerifiedAccounts (profile) {
   return verifs
 }
 
-module.exports = { profileGraphQL, getProfile, getSpace, listSpaces, getThread, getRootStoreAddress, getProfiles, getVerifiedAccounts }
+module.exports = { profileGraphQL, getProfile, getSpace, listSpaces, getThread, getThreadByAddress, getConfig, getRootStoreAddress, getProfiles, getVerifiedAccounts, getSpaceDID }